Clarify ambiguity in 02/Chipotle steps 4 and 5
Update revenue calculation task in 01_Getting_&_Knowing_Your_Data/Chipotle
to be consistent with the understanding that item_price = quantity * price per item.

Clarify step 4 in 02_Filtering_&_Sorting/Chipotle to ask for all items
where the price per item is able to be greater than $10. This is to
accomodate the cases where an item usually costs less than $10, but with
certain choice_description it is able to cost more than $10. To
simplify the task, we simply include such cases and provide the user
with an example.

Similarly, for step 5 in 02_Filtering_&_Sorting/Chipotle, items can
have multiple prices depending on the choice_description they are paired
with. To remove the ambiguity, update the task to ask for the price
of each unique combination of item_name and choice_description.